feat(consumer):
Moditions to allow enable or not polling.

Why?
If I try to using the library on Supabase edge functions or another serverless platform will cause a problem, because the polling will keep the Edge function running for a long time
and serveless solutions are cheap if execute short period of time.

So I resolved using the option enabledPolling equal false where consume the messages and finish the execution.

BREAKING CHANGE: if upgrade for the new version you need to set the enabledPolling on options for true or false
when create a Consumer instance, because is required.

If enabledPolling is true means will keep get message to process, case set as false will consume the message one time and stop.